getToken in helper file
