+++ server/routes/getToken.js
@@ -1,0 +1,13 @@
+//function to get and extract the token from the request headers.
+getToken = function (headers) {
+    if (headers && headers.authorization) {
+        var parted = headers.authorization.split(' ');
+        if (parted.length === 2) {
+            return parted[1];
+        } else {
+            return null;
+        }
+    } else {
+        return null;
+    }
+};
